/src/atoms/auth.ts を整理
