+++ frontend/src/components/AuthCard.tsx
@@ -11,12 +11,6 @@ import {
 import { useAtom } from 'jotai';
 import { useState } from 'react';
 
-export type UserPersonalData = {
-  id: number;
-  email: string;
-  displayName: string;
-};
-
 /**
  * 42認証用のフォーム
  * ボタンが1つあるだけ
@@ -226,7 +220,7 @@ export const DevAuthLoginCard = (props: {
  * 認証済み状態で表示されるUI
  */
 export const DevAuthenticatedCard = (props: { onLogout?: () => void }) => {
-  const [personalData] = useAtom(authAtom.personalDataAtom);
+  const [personalData] = useAtom(authAtom.personalData);
   return (
     <>
       <FTH1 className="text-4xl font-bold" style={{ padding: '4px' }}>
